🚇🧪 Add self consistency test with comparison-results (#3)
This PR adds selfconsistency tests for the original compare results as well as newly created results with pyglotaran main and examples main. This way we ensure that the validation code itself works as expected and keeps working when we update the compare results (e.g. new files are added to the result that might break validation).

It also fixes the crashing validation in `map_result_parameters` because new glotaran results contain `optimization_history.csv` and `parameter_history.csv` which don't have a `label` column

Since the `source_path` is set when loading a dataset with `glotaran.io.load_dataset`, but we read it independently from pyglotaran with `xr.open_dataset` we ignore it when validating.

**note:**
The type ignore is needed because of the difference between numpys 'floating[_32Bit]' and python float.
pyglotaran-examples/test_result_consistency.py:255: error: Incompatible types in assignment (expression has type "float", variable has type "floating[_32Bit]")  [assignment]

* 🩹 Exclude optimization_history.csv and parameter_history.csv from map_result_parameters

* 🚇🧪 Add self consistency test with comparison-results

🚇🧪 Add self consistency test with new results
🚇👌 Run 'Test pyglotaran examples' on schedule

* ⬆️🔧 Update pre-commit config so isort does not crash on installation

* 🩹 Exclude source_path dataset attribute from tests

* 🚇👌 Copy comparison-results to comparison-results-current instead of reclone

* 🚇👌 Copy comparison-results-current to comparison-results for self-consistency test

instead of reclone

* 🧹 Separate steps by a newline (consistency)